+++ tests/Unit/CheckTokenForAnyScopeTest.php
@@ -5,7 +5,7 @@ namespace Laravel\Passport\Tests\Unit;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Laravel\Passport\Exceptions\AuthenticationException;
-use Laravel\Passport\Http\Middleware\CheckClientCredentials;
+use Laravel\Passport\Http\Middleware\CheckTokenForAnyScope;
 use League\OAuth2\Server\Exception\OAuthServerException;
 use League\OAuth2\Server\ResourceServer;
 use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
@@ -13,7 +13,7 @@ use Mockery as m;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ServerRequestInterface;
 
-class CheckClientCredentialsTest extends TestCase
+class CheckTokenForAnyScopeTest extends TestCase
 {
     use MockeryPHPUnitIntegration;
 
@@ -28,7 +28,7 @@ class CheckClientCredentialsTest extends TestCase
             'oauth_scopes' => ['*'],
         ]);
 
-        $middleware = new CheckClientCredentials($resourceServer);
+        $middleware = new CheckTokenForAnyScope($resourceServer);
 
         $request = Request::create('/');
         $request->headers->set('Authorization', 'Bearer token');
@@ -40,7 +40,7 @@ class CheckClientCredentialsTest extends TestCase
         $this->assertSame('response', $response->getContent());
     }
 
-    public function test_request_is_passed_along_if_token_and_scope_are_valid()
+    public function test_request_is_passed_along_if_token_has_any_required_scope()
     {
         $resourceServer = m::mock(ResourceServer::class);
         $resourceServer->shouldReceive('validateAuthenticatedRequest')->andReturn($psr = m::mock(ServerRequestInterface::class));
@@ -48,17 +48,17 @@ class CheckClientCredentialsTest extends TestCase
             'oauth_user_id' => 1,
             'oauth_client_id' => 1,
             'oauth_access_token_id' => 'token',
-            'oauth_scopes' => ['see-profile'],
+            'oauth_scopes' => ['foo', 'bar', 'baz'],
         ]);
 
-        $middleware = new CheckClientCredentials($resourceServer);
+        $middleware = new CheckTokenForAnyScope($resourceServer);
 
         $request = Request::create('/');
         $request->headers->set('Authorization', 'Bearer token');
 
         $response = $middleware->handle($request, function () {
             return new Response('response');
-        }, 'see-profile');
+        }, 'notfoo', 'bar', 'notbaz');
 
         $this->assertSame('response', $response->getContent());
     }
@@ -72,7 +72,7 @@ class CheckClientCredentialsTest extends TestCase
             new OAuthServerException('message', 500, 'error type')
         );
 
-        $middleware = new CheckClientCredentials($resourceServer);
+        $middleware = new CheckTokenForAnyScope($resourceServer);
 
         $request = Request::create('/');
         $request->headers->set('Authorization', 'Bearer token');
@@ -82,7 +82,7 @@ class CheckClientCredentialsTest extends TestCase
         });
     }
 
-    public function test_exception_is_thrown_if_token_does_not_have_required_scopes()
+    public function test_exception_is_thrown_if_token_does_not_have_required_scope()
     {
         $this->expectException('Laravel\Passport\Exceptions\MissingScopeException');
 
@@ -92,16 +92,16 @@ class CheckClientCredentialsTest extends TestCase
             'oauth_user_id' => 1,
             'oauth_client_id' => 1,
             'oauth_access_token_id' => 'token',
-            'oauth_scopes' => ['foo', 'notbar'],
+            'oauth_scopes' => ['foo', 'bar'],
         ]);
 
-        $middleware = new CheckClientCredentials($resourceServer);
+        $middleware = new CheckTokenForAnyScope($resourceServer);
 
         $request = Request::create('/');
         $request->headers->set('Authorization', 'Bearer token');
 
         $response = $middleware->handle($request, function () {
             return 'response';
-        }, 'foo', 'bar');
+        }, 'baz', 'notbar');
     }
 }
