+++ tests/Unit/CheckTokenScopesTest.php
@@ -5,7 +5,7 @@ namespace Laravel\Passport\Tests\Unit;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Laravel\Passport\Exceptions\AuthenticationException;
-use Laravel\Passport\Http\Middleware\CheckClientCredentialsForAnyScope;
+use Laravel\Passport\Http\Middleware\CheckTokenScopes;
 use League\OAuth2\Server\Exception\OAuthServerException;
 use League\OAuth2\Server\ResourceServer;
 use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
@@ -13,7 +13,7 @@ use Mockery as m;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ServerRequestInterface;
 
-class CheckClientCredentialsForAnyScopeTest extends TestCase
+class CheckTokenScopesTest extends TestCase
 {
     use MockeryPHPUnitIntegration;
 
@@ -28,7 +28,7 @@ class CheckClientCredentialsForAnyScopeTest extends TestCase
             'oauth_scopes' => ['*'],
         ]);
 
-        $middleware = new CheckClientCredentialsForAnyScope($resourceServer);
+        $middleware = new CheckTokenScopes($resourceServer);
 
         $request = Request::create('/');
         $request->headers->set('Authorization', 'Bearer token');
@@ -40,7 +40,7 @@ class CheckClientCredentialsForAnyScopeTest extends TestCase
         $this->assertSame('response', $response->getContent());
     }
 
-    public function test_request_is_passed_along_if_token_has_any_required_scope()
+    public function test_request_is_passed_along_if_token_and_scope_are_valid()
     {
         $resourceServer = m::mock(ResourceServer::class);
         $resourceServer->shouldReceive('validateAuthenticatedRequest')->andReturn($psr = m::mock(ServerRequestInterface::class));
@@ -48,17 +48,17 @@ class CheckClientCredentialsForAnyScopeTest extends TestCase
             'oauth_user_id' => 1,
             'oauth_client_id' => 1,
             'oauth_access_token_id' => 'token',
-            'oauth_scopes' => ['foo', 'bar', 'baz'],
+            'oauth_scopes' => ['see-profile'],
         ]);
 
-        $middleware = new CheckClientCredentialsForAnyScope($resourceServer);
+        $middleware = new CheckTokenScopes($resourceServer);
 
         $request = Request::create('/');
         $request->headers->set('Authorization', 'Bearer token');
 
         $response = $middleware->handle($request, function () {
             return new Response('response');
-        }, 'notfoo', 'bar', 'notbaz');
+        }, 'see-profile');
 
         $this->assertSame('response', $response->getContent());
     }
@@ -72,7 +72,7 @@ class CheckClientCredentialsForAnyScopeTest extends TestCase
             new OAuthServerException('message', 500, 'error type')
         );
 
-        $middleware = new CheckClientCredentialsForAnyScope($resourceServer);
+        $middleware = new CheckTokenScopes($resourceServer);
 
         $request = Request::create('/');
         $request->headers->set('Authorization', 'Bearer token');
@@ -82,7 +82,7 @@ class CheckClientCredentialsForAnyScopeTest extends TestCase
         });
     }
 
-    public function test_exception_is_thrown_if_token_does_not_have_required_scope()
+    public function test_exception_is_thrown_if_token_does_not_have_required_scopes()
     {
         $this->expectException('Laravel\Passport\Exceptions\MissingScopeException');
 
@@ -92,16 +92,16 @@ class CheckClientCredentialsForAnyScopeTest extends TestCase
             'oauth_user_id' => 1,
             'oauth_client_id' => 1,
             'oauth_access_token_id' => 'token',
-            'oauth_scopes' => ['foo', 'bar'],
+            'oauth_scopes' => ['foo', 'notbar'],
         ]);
 
-        $middleware = new CheckClientCredentialsForAnyScope($resourceServer);
+        $middleware = new CheckTokenScopes($resourceServer);
 
         $request = Request::create('/');
         $request->headers->set('Authorization', 'Bearer token');
 
         $response = $middleware->handle($request, function () {
             return 'response';
-        }, 'baz', 'notbar');
+        }, 'foo', 'bar');
     }
 }
